% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Main.R
\name{bootstrap_ci}
\alias{bootstrap_ci}
\title{Estimates confidence interval based on bootstrap results}
\usage{
bootstrap_ci(point.est, bootstrap.resamples, alpha = 0.05)
}
\arguments{
\item{point.est}{vector of point estimates (each element represents gene)}

\item{bootstrap.resamples}{matrix of resampled values using bootstrap (each column is a replicate)}

\item{alpha}{significance level}
}
\value{
data frame containing:
    Estimate: point estimates of the statistic
    Lower: lower bound of confidence interval
    Upper: upper bound of confidence interval
}
\description{
Estimates confidence interval based on bootstrap results
}
