% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Main.R
\name{cor2}
\alias{cor2}
\title{Correlation between rows in x vs. y}
\usage{
cor2(x, y, ...)
}
\arguments{
\item{x}{matrix whose rows are to be correlated with y}

\item{y}{vector of numbers}

\item{...}{extra arguments to pass to cor function}
}
\value{
correlation between each row and y
}
\description{
Correlation between rows in x vs. y
}
