% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Main.R
\name{cor_bootstrap}
\alias{cor_bootstrap}
\title{Estimates correlations with y using bootstrapping}
\usage{
cor_bootstrap(x, y, n = 1000, ...)
}
\arguments{
\item{x}{matrix of x values}

\item{y}{y values to be correlated against x}

\item{n}{number of resampling trials}

\item{...}{extra arguments to pass to cor}
}
\value{
resampled correlations where each column represents a resampling
}
\description{
Estimates correlations with y using bootstrapping
}
