% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Main.R
\name{cor_resample}
\alias{cor_resample}
\title{Estimates correlations by taking random subsets of original data and correlating them with y}
\usage{
cor_resample(x, y, frac = 0.5, n = 1000, ...)
}
\arguments{
\item{x}{matrix of x values}

\item{y}{y values to be correlated against x}

\item{frac}{fraction of samples in x to draw}

\item{n}{number of resampling trials}

\item{...}{extra arguments to pass to cor}
}
\value{
resampled correlations
}
\description{
Estimates correlations by taking random subsets of original data and correlating them with y
}
