% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Main.R
\name{cor_test}
\alias{cor_test}
\title{Correlation between rows in x vs. y}
\usage{
cor_test(x, y, n = 1000, ...)
}
\arguments{
\item{x}{matrix whose rows to be correlated with y}

\item{y}{vector of numbers}

\item{n}{number of trials for confidence intervals and P values}

\item{...}{extra arguments to pass to cor function}
}
\value{
data frame of correlations, confidence intervals, P value, false discovery rate
}
\description{
Correlation between rows in x vs. y
}
\examples{
cor_test(t(iris[,1:3]), iris[,4])
}
