% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Main.R
\name{fisher_z}
\alias{fisher_z}
\title{Transform correlation values using Fisher's z transform}
\usage{
fisher_z(r)
}
\arguments{
\item{r}{correlation}
}
\value{
transformed correlation
}
\description{
Transform correlation values using Fisher's z transform
}
