% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Main.R
\name{get_ci}
\alias{get_ci}
\title{Returns confidence interval from bootstrap object}
\usage{
get_ci(b.obj, num.rows)
}
\arguments{
\item{b.obj}{boot object}

\item{num.rows}{how many rows in data originally used to make bootstrap}
}
\value{
endpoints of confidence interval
}
\description{
Returns confidence interval from bootstrap object
}
