% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Main.R
\name{perm_test}
\alias{perm_test}
\title{Tests whether each row in f(x) is equal to 0 using permutation test}
\usage{
perm_test(f, x, x.rand = function() sample(x), n = 1000)
}
\arguments{
\item{f}{function that takes x and produces a vector (must have the same names in same order in every trial!)}

\item{x}{value of x such that f(x) gives observed value}

\item{x.rand}{function that takes no arguments and gives random value of x under null hypothesis}

\item{n}{number of trials}
}
\value{
two-sided P value of observed value compared to values assuming null hypothesis
}
\description{
Tests whether each row in f(x) is equal to 0 using permutation test
}
