% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Main.R
\name{plot_resample}
\alias{plot_resample}
\title{Plots statistics resampled from dataset}
\usage{
plot_resample(results, results2 = NULL)
}
\arguments{
\item{results}{a matrix where each row represents gene, each column represents resampling trial}

\item{results2}{optional second dataset}
}
\value{
ggplot2 graph object
}
\description{
Plots statistics resampled from dataset
}
