% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Main.R
\name{return_null_if_na}
\alias{return_null_if_na}
\title{Given a vector, returns either vector itself or NULL if it has NA values}
\usage{
return_null_if_na(x)
}
\arguments{
\item{x}{value to test}
}
\value{
vector itself if vector doesn't contain NA. NULL otherwise.
}
\description{
Given a vector, returns either vector itself or NULL if it has NA values
}
