% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Main.R
\name{rlm_beta}
\alias{rlm_beta}
\title{Gets the slope of regression of y vs. x}
\usage{
rlm_beta(x, y)
}
\arguments{
\item{x}{matrix where each row is to be regressed vs. y}

\item{y}{vector of numbers}
}
\value{
Coefficients of regression from each row in x vs. y
}
\description{
Gets the slope of regression of y vs. x
}
