% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Main.R
\name{rlm_beta_bootstrap}
\alias{rlm_beta_bootstrap}
\title{Gets bootstrap estimates of slopes of y vs. x}
\usage{
rlm_beta_bootstrap(x, y, n = 1000)
}
\arguments{
\item{x}{matrix where each row is to be regressed vs. y}

\item{y}{vector of numbers (length same as columns of x)}

\item{n}{number of trials}
}
\value{
matrix where each row represents slopes for each row in x, each column represents a trial
}
\description{
Gets bootstrap estimates of slopes of y vs. x
}
