% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Main.R
\name{rlm_beta_diff}
\alias{rlm_beta_diff}
\title{Returns difference in the regression coefficients of y vs. x in two sample classes}
\usage{
rlm_beta_diff(x, y, x.class)
}
\arguments{
\item{x}{matrix of values whose rows are to be regressed vs. y. Columns represent samples.}

\item{y}{vector of numbers to regress against}

\item{x.class}{vector of TRUE or FALSE values indicating group assignment of each sample in x.}
}
\value{
vector of difference of regression coefficients for each row in x. Sign convention: TRUE group - FALSE group.
}
\description{
Returns difference in the regression coefficients of y vs. x in two sample classes
}
