% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Main.R
\name{rlm_beta_diff_test}
\alias{rlm_beta_diff_test}
\title{Test whether there is difference in slopes of y vs. x in two classes of samples}
\usage{
rlm_beta_diff_test(x, y, x.class, n = 1000)
}
\arguments{
\item{x}{matrix where rows = genes, columns = samples}

\item{y}{value to be regressed with each row}

\item{x.class}{boolean vector (TRUE or FALSE) for each sample indicating sample group}

\item{n}{number of permutations}
}
\value{
data frame of estimates, confidence intervals, P values for slopes in each sample class
    as well as the difference between the classes.
}
\description{
Test whether there is difference in slopes of y vs. x in two classes of samples
}
\note{
In the output, Class1 refers to the TRUE group while Class2 refers to the FALSE group.
}
