% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Main.R
\name{rlm_beta_test}
\alias{rlm_beta_test}
\title{Fits linear model to predict y using each row in x and tests whether each slope is significant.}
\usage{
rlm_beta_test(x, y, n = 1000)
}
\arguments{
\item{x}{matrix whose rows are to be regressed vs. y}

\item{y}{vector of numbers}

\item{n}{number of trials for bootstrap and confidence intervals}
}
\value{
slopes of each row in x vs. y, confidence intervals, and P values
}
\description{
Fits linear model to predict y using each row in x and tests whether each slope is significant.
}
\examples{
rlm_beta_test(t(iris[,1:3]), iris[,4], n=100)
}
