% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Main.R
\name{rlm_coef}
\alias{rlm_coef}
\title{Fits y to x using robust linear regression and returns regression coefficients}
\usage{
rlm_coef(x, y)
}
\arguments{
\item{x}{matrix of values where rows = genes, columns = samples}

\item{y}{value for every sample in x}
}
\value{
coefficients from robust linear regression
}
\description{
Fits y to x using robust linear regression and returns regression coefficients
}
\examples{
rlm_coef(t(iris[,1:3]), iris[,4])
}
