% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Main.R
\name{rlm_coef_bootstrap}
\alias{rlm_coef_bootstrap}
\title{Returns robust linear regression coefficients by bootstrapping}
\usage{
rlm_coef_bootstrap(x, y, n = 1000)
}
\arguments{
\item{x}{dataset where rows represent genes, columns represent samples}

\item{y}{values for each sample in x}

\item{n}{number of resamples}
}
\value{
matrix of resamples
}
\description{
Returns robust linear regression coefficients by bootstrapping
}
