% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Main.R
\name{rlm_coef_resample}
\alias{rlm_coef_resample}
\title{Returns robust linear regression coefficients by resampling}
\usage{
rlm_coef_resample(x, y, frac = 0.5, n = 1000)
}
\arguments{
\item{x}{dataset where rows represent genes, columns represent samples}

\item{y}{values for each sample in x}

\item{frac}{fraction of samples to use in resample}

\item{n}{number of resamples}
}
\value{
matrix of resamples
}
\description{
Returns robust linear regression coefficients by resampling
}
