% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Main.R
\name{two_cor}
\alias{two_cor}
\title{Returns difference between the Fisher transformed correlations between x and y
    in two classes of samples}
\usage{
two_cor(x, y, x.classes, ...)
}
\arguments{
\item{x}{matrix of x values (rows = genes, columns = samples)}

\item{y}{y values to correlate against (same length as number of samples in x)}

\item{x.classes}{vector of TRUE and FALSE values for each sample in x, respectively}

\item{...}{extra parameters to pass to cor2}
}
\value{
vector of difference between correlations in two sets of samples (TRUE group - FALSE group)
}
\description{
Returns difference between the Fisher transformed correlations between x and y
    in two classes of samples
}
