% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Main.R
\name{two_cor_bootstrap}
\alias{two_cor_bootstrap}
\title{Bootstrap estimates of difference between two Fisher z-transformed correlations}
\usage{
two_cor_bootstrap(x, y, x.classes, n = 1000, ...)
}
\arguments{
\item{x}{matrix of x values (rows = genes, columns = samples)}

\item{y}{values to correlate against (must be same length as number of samples in x)}

\item{x.classes}{vector of TRUE of FALSE values for every sample in x}

\item{n}{number of trials}

\item{...}{parameters to pass to cor2}
}
\value{
matrix of bootstrap estimates of differences in correlations.
    Sign convention is TRUE group - FALSE group.
    Each column represents bootstrap iteration
}
\description{
Bootstrap estimates of difference between two Fisher z-transformed correlations
}
