% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Main.R
\name{two_cor_test}
\alias{two_cor_test}
\title{Tests hypothesis that correlation of x with y differs in two sample classes using permutation test}
\usage{
two_cor_test(x, y, x.classes, n = 1000, ...)
}
\arguments{
\item{x}{matrix of x values (rows = genes, columns = samples)}

\item{y}{y values to correlate against (must be same length as number of samples in x)}

\item{x.classes}{a vector of TRUE and FALSE values for every sample in x, respectively}

\item{n}{number of trials for permutation test}

\item{...}{extra arguments to pass into cor}
}
\value{
data frame with components:
Cor1 (correlations in samples marked TRUE),
Cor2 (correlations in samples marked FALSE)
P value that difference in correlations is significant
}
\description{
Tests hypothesis that correlation of x with y differs in two sample classes using permutation test
}
