% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Main.R
\name{two_rlm_beta_test}
\alias{two_rlm_beta_test}
\title{Tests whether the slopes of y vs. x is the same in two sample groups}
\usage{
two_rlm_beta_test(x, y, x.class, n = 1000)
}
\arguments{
\item{x}{matrix where each column contains a sample and each row is to be regressed vs. y}

\item{y}{values to be regressed against}

\item{x.class}{class of each sample in x, respectively (can be TRUE or FALSE for each sample)}

\item{n}{number of permutations}
}
\value{
Data frame of bootstrap estimates for each sample group and P value of difference
    between two sample groups
}
\description{
Tests whether the slopes of y vs. x is the same in two sample groups
}
