% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Main.R
\name{two_rlm_coef_test}
\alias{two_rlm_coef_test}
\title{Compares coefficients from two robust linear models using permutation test}
\usage{
two_rlm_coef_test(x1, x2, y, n = 1000)
}
\arguments{
\item{x1}{matrix where rows represent predictors and columns represent samples}

\item{x2}{exactly like x1 but for a separate model}

\item{y}{value to predict for each sample}

\item{n}{number of permutations}
}
\value{
data frame of
   Coef1 the coefficients from first model
   Coef2 the coefficients from the second model
   P P value that coefficients differ between two models
}
\description{
Compares coefficients from two robust linear models using permutation test
}
