% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Main.R
\name{cor2}
\alias{cor2}
\title{Computes the correlation of each gene with vector of y values}
\usage{
cor2(x, y, ...)
}
\arguments{
\item{x}{matrix of x values (rows = genes, columns = samples)}

\item{y}{y values (vector with as many samples as x)}

\item{...}{extra arguments to pass into correlation function}
}
\value{
matrix of correlations where each row represents correlation of each gene with y
}
\description{
Computes the correlation of each gene with vector of y values
}
