% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Main.R
\name{cor_test}
\alias{cor_test}
\title{Permutation test on correlation between x and y}
\usage{
cor_test(x, y, n = 1000, ...)
}
\arguments{
\item{x}{matrix of x values (rows = genes, columns = samples)}

\item{y}{vector of y values to be correlated with x}

\item{n}{number of trials for permutation test}

\item{...}{extra parameters to pass into cor function}
}
\value{
data frame with components:
   Cor (correlation of each row in x with y)
   P (P value that correlation is significantly different from 0)
}
\description{
Permutation test on correlation between x and y
}
