% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Main.R
\name{rlm_beta}
\alias{rlm_beta}
\title{Fits linear model to predict y using each row in x one at a time. Returns coefficients of fit.}
\usage{
rlm_beta(x, y)
}
\arguments{
\item{x}{matrix whose rows are to be regressed vs. y}

\item{y}{vector of numbers}
}
\value{
slopes of the robust linear model fits of y vs. each variable in x
}
\description{
Fits linear model to predict y using each row in x one at a time. Returns coefficients of fit.
}
