% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Main.R
\name{rlm_beta_test}
\alias{rlm_beta_test}
\title{Tests whether slopes of linear regressions equal 0 using permutation test}
\usage{
rlm_beta_test(x, y, n = 1000)
}
\arguments{
\item{x}{matrix where each row is to be regressed vs. y}

\item{y}{vector of numbers (length same as columns in x)}

\item{n}{number of trials}
}
\value{
Data frame of observed value, upper and lower 95% confidence intervals, and P values
}
\description{
Tests whether slopes of linear regressions equal 0 using permutation test
}
