% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Main.R
\name{rlm_coef_test}
\alias{rlm_coef_test}
\title{Tests hypothesis that the coefficients for robust linear regression != 0 using permutation test}
\usage{
rlm_coef_test(x, y, n = 1000)
}
\arguments{
\item{x}{matrix where genes in rows, samples in columns}

\item{y}{value for every sample for regression}

\item{n}{number of trials for permutation test}
}
\value{
data frame containing
   Coef coefficients from linear regression
   P P values that different from 0
}
\description{
Tests hypothesis that the coefficients for robust linear regression != 0 using permutation test
}
